test and display api data

+++ lib/core/models/rating.dart
@@ -5,7 +5,7 @@ class Rating {
   Rating({required this.rate, required this.count});
 
   Rating.fromJson(Map<String, dynamic> json) {
-    rate = json["rate"];
+    rate = json["rate"].toDouble();
     count = json["count"];
   }
